format and structure changes
