+++ commands/build.js
@@ -1,7 +1,7 @@
 import fs from 'fs'
 import { Clack } from '@archetype-org/clack'
-import { copyDeskToUrbit }  from "../lib/files.js"
-import { isInGolemProject, isUrbitInstalled, isShipCreated, isDeskMountedOnShip } from "../lib/checks.js"
+import { copyDeskToUrbit }  from '../lib/files.js'
+import { isInGolemProject, isUrbitInstalled, isShipCreated, isDeskMountedOnShip } from '../lib/checks.js'
 
 async function build () {
   try {
