+++ lib/accounts.js
@@ -1,20 +1,20 @@
-import { Account } from "@near-js/accounts"
-import { connect, keyStores } from "near-api-js" 
+import { Account } from '@near-js/accounts'
+import { connect, keyStores } from 'near-api-js' 
 import { homedir } from 'os'
 import * as pathlib from 'path'
 
 const homedDirectory = homedir()
-const CREDENTIALS_DIR = ".near-credentials"
+const CREDENTIALS_DIR = '.near-credentials'
 const credentialsPath = pathlib.join(homedDirectory, CREDENTIALS_DIR)
 const myKeyStore = new keyStores.UnencryptedFileSystemKeyStore(credentialsPath)
 
 const connectionConfig = {
-  networkId: "testnet",
+  networkId: 'testnet',
   keyStore: myKeyStore,
-  nodeUrl: "https://rpc.testnet.near.org",
-  walletUrl: "https://testnet.mynearwallet.com/",
-  helperUrl: "https://helper.testnet.near.org",
-  explorerUrl: "https://testnet.nearblocks.io",
+  nodeUrl: 'https://rpc.testnet.near.org',
+  walletUrl: 'https://testnet.mynearwallet.com/',
+  helperUrl: 'https://helper.testnet.near.org',
+  explorerUrl: 'https://testnet.nearblocks.io',
 }
 
 async function getAccount () {
