+++ lib/checks.js
@@ -1,6 +1,6 @@
 import { Clack } from '@archetype-org/clack'
-import { fileExists, folderExists } from "./files.js"
-import { fetchUrbitBinary, bootFakeShip, restartFakeShip } from "./urbit.js"
+import { fileExists, folderExists } from './files.js'
+import { fetchUrbitBinary, bootFakeShip, restartFakeShip } from './urbit.js'
 
 async function isInGolemProject () {
   const condition = await fileExists('ships/ships.json')
