+++ lib/constants.js
@@ -1,0 +1,5 @@
+const CONTRACT_ID = 'lamentable-hobbies.testnet'
+
+export {
+	CONTRACT_ID,
+}
