+++ lib/files.js
@@ -64,6 +64,7 @@ EOF
 
 async function copyDeskToUrbit (desk, ship) {
   try {
+    // todo: make dynamic
     await execShPromise(`cp -r ./apps/${desk}/desk/* ./ships/${ship}/*`, true)
     await execShPromise(`cp -r ./apps/${desk}/desk-deps/urbit/* ./ships/${ship}/*`, true)
     await execShPromise(`cp -r ./apps/${desk}/desk-deps/landscape/* ./ships/${ship}/*`, true)
@@ -92,6 +93,29 @@ function isFolder (path) {
   return !extname(path)
 }
 
+function packageLibrary (path) {
+  // local checks & conditionals
+  const isContainedIn = (array, item) => new Set(array).has(item)
+  const isInSpecialFolder = (path) => isContainedIn(
+    ['lib', 'mar', 'sur', 'ted', 'app'],
+    path.split('/')[0]
+  )
+  const isRoot = (path) => path[0] === '/'
+
+  // get files at path
+  const projectName = basename(process.cwd())
+  const libraryPath = isRoot(path) ? `./apps/${projectName}/desk${path}` :
+    isInSpecialFolder(path) ?  `./apps/${projectName}/desk/${path}` :
+      `./apps/${projectName}/desk/lib/${path}`
+  const files = getFileList(libraryPath) 
+  
+  return {
+    // todo: include accountname
+    package: basename(libraryPath),
+    files, 
+  }
+}
+
 export {
   fileExists,
   folderExists,
@@ -101,4 +125,5 @@ export {
   copyDeskToUrbit,
   getFileList,
   isFolder,
+  packageLibrary,
 }
