+++ lib/urbit.js
@@ -1,4 +1,4 @@
-import execSh from "exec-sh"
+import execSh from 'exec-sh'
 const execShPromise = execSh.promise
 async function fetchUrbitBinary () {
   try {
