+++ lib/utils.js
@@ -1,8 +1,8 @@
 // casting utils
 
 function truthyString (s) {
-  if (typeof s === "boolean") return s
-  if (typeof s !== "string") throw new Error(`truthyString: provided value ${s} is not a string`)
+  if (typeof s === 'boolean') return s
+  if (typeof s !== 'string') throw new Error(`truthyString: provided value ${s} is not a string`)
   return s === 'true'
 }
