+++ templates/sail-desk.js
@@ -1,7 +1,7 @@
 import { cruft } from '../partials/cruft.js'
 import { sailFiles } from '../partials/sail-files.js'
 import { itemFiles } from '../partials/item-files.js'
-import { reactFiles } from "../partials/react-files.js"
+import { reactFiles } from '../partials/react-files.js'
 
 function sailDesk (shipName, deskName) {
   const files = [
